fix scoping on ionAudioPlay directive
The ionAudioPlay directive has a scope watcher for `track.status` which
it uses to change the Play/Pause button when a track has stopped (either
because it finished playing, or the file was not found)

However the directive didn't have the track in scope (looks like it used
to, but the feature got lost at some point?) so the button never changes

This follows grabs the track from the ionAudioTrack parent directive, and
places it in it's own scope. Just like how ionAudioControls already does.